(#64) Move error message extractors to API specific routines
Though it is not the best solution, this is what works and does the
job. For now.

Fixes #64
